more pulling apart and putting back together
